Folder Creation Fix/Batch Mode Set
Now the script will create the specified folders if they do not already exist and if the user wishes to use folders. If the user does not wish to use folders, It will just ensure the filepath has been created.

Setting Batch Mode has been added as well, but it does not seem to work. When I run the script with Batch Mode both On and Off, the export/save dialogue boxes still appear. Looking into it.